Joomla 1.0.10

+++ components/components/com_weblinks/weblinks.php
@@ -1,6 +1,6 @@
 <?php
 /**
-* @version $Id: weblinks.php 3594 2006-05-22 17:29:07Z stingrey $
+* @version $Id: weblinks.php 4060 2006-06-19 22:31:10Z stingrey $
 * @package Joomla
 * @subpackage Weblinks
 * @copyright Copyright (C) 2005 Open Source Matters. All rights reserved.
@@ -284,7 +284,6 @@ function cancelWebLink( $option ) {
 * @param database A database connector object
 */
 function saveWeblink( $option ) {
-	global $mosConfig_mailfrom, $mosConfig_fromname;
 	global $database, $my;
 
 	if ($my->gid < 1) {
@@ -292,6 +291,23 @@ function saveWeblink( $option ) {
 		return;
 	}
 
+	// security check to see if link exists in a menu
+	$link = 'index.php?option=com_weblinks&task=new';
+	$query = "SELECT id"
+	. "\n FROM #__menu"
+	. "\n WHERE link LIKE '%$link%'"
+	. "\n AND published = 1"
+	;
+	$database->setQuery( $query );
+	$exists = $database->loadResult();
+	if ( !$exists ) {						
+		mosNotAuth();
+		return;
+	}		
+	
+	// simple spoof check security
+	josSpoofCheck();	
+	
 	$row = new mosWeblink( $database );
 	if (!$row->bind( $_POST, 'published' )) {
 		echo "<script> alert('".$row->getError()."'); window.history.go(-1); </script>\n";
@@ -302,8 +318,8 @@ function saveWeblink( $option ) {
 	$row->date = date( 'Y-m-d H:i:s' );
 	
 	// until full edit capabilities are given for weblinks - limit saving to new weblinks only
-	$row->id = 0;
-
+	$row->id = 0;	
+	
 	if (!$row->check()) {
 		echo "<script> alert('".$row->getError()."'); window.history.go(-1); </script>\n";
 		exit();
